+++ packages/dappmanager/src/modules/optimism/index.ts
@@ -1,0 +1,2 @@
+export { getOptimismConfig } from "./getOptimismConfig.js";
+export { setOptimismConfig } from "./setOptimismConfig.js";
