+++ packages/dappmanager/src/modules/optimism/params.ts
@@ -1,0 +1,4 @@
+
+
+export const opNodeRpcUrlEnvName = "L1_RPC";
+export const opNodeServiceName = "op-node";
